{topost} net: hns3: fix query fd information missing issue
In oringinal codes, the flow_type of fd rule was broken, the flags
FLOW_EXT and FLOW_MAC_EXT were dropped. When query the fd information,
the extra information of the rules were lost. The mac information of
FLOW_MAC_EXT is also incorrect by using the mask value. This patch fixes
them.

Fixes: 80f70308e381 ("{topost} net: hns3: add support for rule add/delete for flow director")

+++ drivers/net/ethernet/hisilicon/hns3/hns3pf/hclge_main.h
@@ -417,7 +417,7 @@ struct hclge_fd_rule {
 	struct hclge_fd_rule_tuples tuples;
 	struct hclge_fd_rule_tuples tuples_mask;
 	u32 unused_tuple;
-	u8 flow_type;
+	u32 flow_type;
 	u8 action;
 	u16 vf_id;
 	u16 queue_id;
